feat: Added Items - pellets and energizers - and win condition
Added an Item class for items
Items are always contained in a Tile which is responsible for rendering them (and maybe updating them in the furute?)
The Map has a vector of all items and is responsible for creating and deleting them.

When Pac-Man eats a Pellet he gets 10 score and is slowed down by one frame.
When Pac-Man eats an Energizer he gets 50 score and is slowed down by three frames.