utility: Added Timer class and utility method to get direction from orientation
Added Timer class that makes it easier to check when a certain amount of time has passed.
Can be used for animations and temporary effects.

Direction Utilities::GetDirectionFromOrientation(Orientation orientation);
Returns a direction based on an orientation.
Can be used to avoid receiving an undesired Direction::NONE for a stationary target.