feat: Inky now targets as in the original game
Inky mirrors Blinky's position using two tiles ahead of Pac-Man as the origin point.
He takes the tile distance and the direction between Blinky and two tiles ahead of Pac-Man and doubles the distance in the same direction. The resulting tile is his target.
If two tiles ahead of Pac-Man is outside of the map he uses Pac-Man's tile instead to calculate the target tile.
If his target tile is outside of the map he targets Pac-Man's tile directly.

Also fixed Utilities::GetDirectionFromOrientation() which was returning the opposite of the expected direction.